+++ src/constants/date-locales.js
@@ -1,3 +1,0 @@
-export const LOCALES = Object.freeze({
-  // here will be locales as values and their names as keys
-})