+++ src/i18n/index.js
@@ -8,13 +8,13 @@ import i18nextBrowserLanguageDetector from 'i18next-browser-languagedetector'
 // https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
 
 // TODO: make i18n.language and i18n.languages responsive
-import moment from 'moment-timezone'
 
 import _isObject from 'lodash/isObject'
 import _get from 'lodash/get'
 import _merge from 'lodash/merge'
 
 import { MathUtil } from '@/utils/math.util'
+import { DateUtil } from '@/utils/date.util'
 
 class I18n {
   constructor () {
@@ -158,20 +158,19 @@ class I18n {
 
           switch (format.toLowerCase()) {
             case 'date':
-              return moment(param)
-                .format(_get(lngConfig, 'date.presets.datetime'))
+              return DateUtil.format(param, _get(lngConfig, 'date.presets.datetime'))
             case 'dmy':
-              return moment(param)
-                .format(_get(lngConfig, 'date.presets.dmy'))
+              return DateUtil.format(param, _get(lngConfig, 'date.presets.dmy'))
             case 'dmyt':
-              return moment(param)
-                .format(_get(lngConfig, 'date.presets.dmyt'))
+              return DateUtil.format(param, _get(lngConfig, 'date.presets.dmyt'))
             case 'calendar':
-              return moment(param)
-                .calendar(null, _get(lngConfig, 'date.calendar'))
+              return DateUtil.toHuman(
+                param, undefined, _get(lngConfig, 'date.calendar')
+              )
             case 'calendar-inline':
-              return moment(param)
-                .calendar(null, _get(lngConfig, 'date.calendarInline'))
+              return DateUtil.toHuman(
+                param, undefined, _get(lngConfig, 'date.calendarInline')
+              )
             case 'money':
               const value = (_isObject(param) ? param.value : param) || '0'
               const defaultFormat =